persist: add fetch submodule to read
